+++ CustomRoles.cs
@@ -32,7 +32,7 @@ public enum CustomRoles
     BountyHunter,
     Butcher,
     Camouflager,
-    Capitalism,
+    Capitalist,
     Cantankerous,
     Catalyst,
     Centralizer,
@@ -234,7 +234,6 @@ public enum CustomRoles
     Mortician,
     NiceEraser,
     NiceGuesser,
-    NiceSwapper,
     Nightmare,
     Negotiator,
     Observer,
@@ -266,6 +265,7 @@ public enum CustomRoles
     SpeedBooster,
     Spy,
     SuperStar,
+    Swapper,
     TaskManager,
     Telekinetic,
     Tether,
