+++ Patches/ChatCommandPatch.cs
@@ -391,7 +391,7 @@ internal static class ChatCommands
         {
             if (GuessManager.GuesserMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
             if (Judge.TrialMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
-            if (NiceSwapper.SwapMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
+            if (Swapper.SwapMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
             if (Inspector.InspectorCheckMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
             if (Councillor.MurderMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
             if (Medium.MsMsg(PlayerControl.LocalPlayer, text)) goto Canceled;
@@ -3595,7 +3595,7 @@ internal static class ChatCommands
         {
             if (GuessManager.GuesserMsg(player, text) ||
                 Judge.TrialMsg(player, text) ||
-                NiceSwapper.SwapMsg(player, text) ||
+                Swapper.SwapMsg(player, text) ||
                 Inspector.InspectorCheckMsg(player, text) ||
                 Councillor.MurderMsg(player, text))
             {
