+++ Resources/Lang/en_US.json
@@ -210,7 +210,7 @@
   "Bomber": "Bomber",
   "Scavenger": "Scavenger",
   "Trapster": "Trapster",
-  "Capitalism": "Capitalist",
+  "Capitalist": "Capitalist",
   "Gangster": "Gangster",
   "Cleaner": "Cleaner",
   "Lightning": "Lightning",
@@ -489,7 +489,7 @@
   "TrapsterInfo": "Start a killing chain",
   "ScavengerInfo": "Your kills are unreportable",
   "EvilGuesserInfo": "Guess roles in meetings to kill",
-  "CapitalismInfo": "Give more tasks to crewmates",
+  "CapitalistInfo": "Give more tasks to crewmates",
   "GangsterInfo": "Convert players to your side",
   "CleanerInfo": "Report bodies to make them unreportable",
   "LightningInfo": "Convert players to Quantum Ghosts",
@@ -719,7 +719,7 @@
   "BomberInfoLong": "(Impostors):\\nThe Bomber can use the vanish/pet button to self-explode, killing players within a certain range. But as a price, the Bomber will also die. Note: All players will see a kill flash when the Bomber explodes.",
   "ScavengerInfoLong": "(Impostors):\\nScavenger kills do not leave dead bodies behind. To kill someone, you must click your kill button on them, then stay near that person for a few seconds, similar to an Arsonist douse.",
   "TrapsterInfoLong": "(Impostors):\\nAs the Trapster, you can vanish/pet to create a fake dead body. Anyone who reports this body will die, and depending on settings the body will be reported.\\n\\n(Legacy Version):\\nIn the game, the Trapster kills a player and then puts a trap on their body. If another player reports the dead body and it has the Trapster's trap on it, the player who reports will die instantly without the body getting reported. The player killed by the Trapster's trap will also become a Trapster. This chain reaction continues until no more players are left to become Trapsters. However, the Trapster will die immediately after making a kill. It's worth noting that the Trapster's own dead body cannot trigger their trap.",
-  "CapitalismInfoLong": "(Impostors):\\nCapitalists can add a short task to a player by pressing the kill button. Note: using this ability on players who have no tasks or have completed tasks will not take effect.",
+  "CapitalistInfoLong": "(Impostors):\\nCapitalists can add a short task to a player by pressing the kill button. Note: using this ability on players who have no tasks or have completed tasks will not take effect.",
   "GangsterInfoLong": "(Impostors):\\nThe Gangster can recruit a player to a Madmate by pressing the kill button. If successful, both will see the shield animation as a reminder. The host sets the maximum number of recruits. If recruitment fails, the target is killed normally. When out of recruits, the Gangster can only make normal kills.",
   "CleanerInfoLong": "(Impostors):\\nCleaner can press the Report button to clean up any dead body they come across.\\n\\nNote: You can only clean bodies if your kill cooldown is up.\\n\\nIf the cleanup is successful, the Cleaner will see a shield animation on their body as a reminder (only visible to himself). The cleaned up body cannot be reported (including Bait's).",
   "LightningInfoLong": "(Impostors):\\nThe Lightning role cannot make regular kills. When the Lightning uses its kill button, the target will vaporize into a quantum ghost. If a player who is not a quantum ghost touches the quantum ghost, the quantum ghost will die. If the quantum ghost survives until a meeting, they will die after the meeting ends. Interacting with the quantum ghost has no effect. Killing the Lightning may cause the killer to vaporize into a quantum ghost, depending on the host's settings.",
@@ -1609,7 +1609,7 @@
   "SnitchCanBeWorkhorse": "<color=#b8fb4f>Snitch</color> can become <color=#00ffff>Workhorse</color>",
   "WorkhorseSpawnChance": "<color=#00ffff>Workhorse</color> Spawn Chance",
   "InnocentCanWinByImp": "If their target was an <color=#ff1919>Impostor</color> then they win with them",
-  "CapitalismSkillCooldown": "Ability Cooldown",
+  "CapitalistSkillCooldown": "Ability Cooldown",
   "DualVotes": "Duplicate votes",
   "HideDualVotes": "Hide duplicated votes",
   "VeteranSkillCooldown": "Alert Cooldown",
@@ -1845,7 +1845,7 @@
   "CheckReportFail": "<#ff0000>This meeting call was canceled due to the following reason:\\n</color>",
   "CannotCallEmergencyMeetingWhileSabotage": "Emergency meetings cannot be called during crisis",
   "JesterCannotCallEmergencyMeeting": "As the <#ec62a5>Jester</color>, according to the host's settings, you cannot call emergency meetings",
-  "NiceSwapperCannotCallEmergencyMeeting": "As the <#922348>Nice Swapper</color>, according to the host's settings, you cannot call emergency meetings",
+  "SwapperCannotCallEmergencyMeeting": "As the <#922348>Nice Swapper</color>, according to the host's settings, you cannot call emergency meetings",
   "AdrenalineCannotCallEmergencyMeeting": "As the <#ffff00>Adrenaline</color>, you cannot call emergency meetings currently",
   "PlayerWasRevived": "The player whose body you tried to report was revived",
   "HypnosisNoMeeting": "Everyone is in hypnosis, no meeting calls are allowed currently",
@@ -3259,7 +3259,7 @@
   "BountyHunterChangeButtonText": "Swap",
   "EvilTrackerChangeButtonText": "Track",
   "InnocentButtonText": "Frame",
-  "CapitalismButtonText": "Add Task",
+  "CapitalistButtonText": "Add Task",
   "PelicanButtonText": "Eat",
   "CounterfeiterButtonText": "Cheat",
   "PursuerButtonText": "Trick",
@@ -4791,11 +4791,11 @@
   "ExpressInfoLong": "(Crewmates):\\nAs the Express, you gain a small temporary speed boost after each task you complete.",
   "ExpressSpeed": "Increased Speed",
   "ExpressSpeedDur": "Increased Speed Duration",
-  "NiceSwapper": "Swapper",
-  "NiceSwapperInfo": "Swap players' votes",
-  "NiceSwapperInfoLong": "(Crewmates):\\nAs the Swapper, you can use the command <b>/sw <color=#0000ff>[ID]</color></b> twice to choose 2 players. Those votes received by those 2 players will be swapped.\\nYou can see everyone's <color=#0000ff>ID</color>s next to their name.\\n\\nFor example, if 3 players voted red, and 5 players voted yellow, and you swap red and yellows' IDs, then red will receive 5 votes, and yellow will receive 3 votes.\\n\\nDue to technical difficulties, the Host's original vote will seem unchanged, but it's only visual, their vote will still get swapped.",
-  "NiceSwapperMax": "Initial Number of Swaps",
-  "NiceSwapperTrialMax": "Ability use limit reached",
+  "Swapper": "Swapper",
+  "SwapperInfo": "Swap players' votes",
+  "SwapperInfoLong": "(Crewmates):\\nAs the Swapper, you can use the command <b>/sw <color=#0000ff>[ID]</color></b> twice to choose 2 players. Those votes received by those 2 players will be swapped.\\nYou can see everyone's <color=#0000ff>ID</color>s next to their name.\\n\\nFor example, if 3 players voted red, and 5 players voted yellow, and you swap red and yellows' IDs, then red will receive 5 votes, and yellow will receive 3 votes.\\n\\nDue to technical difficulties, the Host's original vote will seem unchanged, but it's only visual, their vote will still get swapped.",
+  "SwapperMax": "Initial Number of Swaps",
+  "SwapperTrialMax": "Ability use limit reached",
   "CanSwapSelfVotes": "<color=#922348>Swapper</color> Can Swap Themselves",
   "SwapperHideMsg": "<color=#00ff00>Hide <color=#922348>Swapper</color>'s commands</color>",
   "CantSwapSelf": "You can't swap yourself",
